+++ src/pages/AdminLogin.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import ErrorNotification from './ErrorNotification';
+import ErrorNotification from '../components/ErrorNotification';
 
 const AdminLogin = () => {
   const navigate = useNavigate();
