+++ src/pages/ArticleManagement.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { Plus, Search, Edit, Trash2 } from 'lucide-react';
 import axios from 'axios';
-import DeleteConfirmationModal from './DeleteConfirmationModal';
+import DeleteConfirmationModal from '../components/DeleteConfirmationModal';
 
 const ArticleManagement = () => {
   const navigate = useNavigate();
