+++ src/pages/CategoryManagement.jsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { Plus, Search, Edit, Trash2 } from 'lucide-react';
-import DeleteConfirmationModal from './DeleteConfirmationModal';
-import SuccessNotification from './SuccessNotification';
+import DeleteConfirmationModal from '../components/DeleteConfirmationModal';
+import SuccessNotification from '../components/SuccessNotification';
 
 const CategoryManagement = () => {
   const navigate = useNavigate();
